--- Dutify/src/components/carousel/old_carousel.css
@@ -1,90 +1,0 @@
-.carrusel-container {
-    background-color: var(--color-background);
-    width: 100%;
-    display: flex;
-    align-items: center;
-    flex-direction:column;
-    gap: 20px;
-}
-
-.carrusel-box {
-    display: flex;
-    align-items: flex-start;
-    flex-direction:column;
-    gap:10px;
-}
-
-.carrusel-h5 {
-    color: var(--color-text);
-    font-size: 25px;
-}
-
-.carrusel-caja-texto {
-    background-color: rgba(0, 0, 0, 0.4);
-    width: 30vw;
-    max-height: 40px;
-    height: calc(20vw * (200 / 800));
-    max-width: 300px;
-    text-align: center;
-    left: 32%;
-    bottom: 0rem;
-    padding: 0;
-    border-top-left-radius: 4px;
-    border-top-right-radius: 4px;
-}
-
-@media (min-width: 701px) {
-    .carrusel-caja-texto {
-        font-size: 1.2rem;
-    }
-}
-
-.carrusel-texto {
-    display: inline-block;
-    max-width: 100%;
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    color: rgba(255, 255, 255, 0.9);
-    margin: 0;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-
-.carousel-item img {
-    object-fit: cover;
-    width: 95vw;
-    height: calc(95vw * (250 / 875));
-    max-width: 875px;
-    max-height: 250px;
-    border-radius: 5px;
-    cursor: pointer;
-}
-  
-.button-svg {
-    width: 70%;
-    height: 70%;
-}
-
-.carousel .carousel-button {
-    opacity: 1;
-}
-
-.carousel {
-    width: 95vw;
-    height: calc(95vw * (250 / 875));
-    max-width: 875px;
-    max-height: 250px;
-    border-radius: 15px;
-}
-
-.carousel button:hover {
-    opacity: 0.6;
-}
-
-
-  
-
