--- Dutify/src/components/carousel/prueba.jsx
@@ -1,75 +1,0 @@
-import React, { useState } from 'react';
-import Slider from 'react-slick';
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
-import './CarouselComponent.css';
-
-const CarouselComponent = ({ lista, name, id }) => {
-  const [activeIndex, setActiveIndex] = useState(0);
-
-  const settings = {
-    dots: true,
-    infinite: true,
-    speed: 500,
-    slidesToShow: 3,
-    slidesToScroll: 1,
-    autoplay: false,
-    autoplaySpeed: 2000,
-    cssEase: "linear",
-    centerMode: true,
-    centerPadding: '0',
-    beforeChange: (oldIndex, newIndex) => handleChange(newIndex),
-  };
-
-  const ClickHandler = (id) => {
-    window.location.href = "/listas/playlist?playlistId=" + id;
-  };
-
-  const handleKeyDown = (event, id) => {
-    if (event.key === 'Enter') {
-      ClickHandler(id);
-    }
-  };
-
-  const handleChange = (index) => {
-    setActiveIndex(index);
-    // Anunciar el nombre de la playlist cuando cambia
-    const activePlaylist = lista[index];
-    if (activePlaylist) {
-      const announcement = document.getElementById(`playlist-announce-${id}`);
-      if (announcement) {
-        announcement.textContent = `Playlist actual: ${activePlaylist.description}`;
-      }
-    }
-  };
-
-  return (
-    <section className='carrusel-container' aria-label={name} id={id}>
-      <div className='carrusel-box'>
-        <h5 className='carrusel-h5' tabIndex="0" aria-labelledby={id}>{name}</h5>
-        <div className="carousel-container">
-          <Slider {...settings}>
-            {lista.map((playlist, index) => (
-              <div
-                key={playlist.id}
-                className="carousel__item"
-                tabIndex="0"
-                role="button"
-                onClick={() => ClickHandler(playlist.id)}
-                onKeyDown={(event) => handleKeyDown(event, playlist.id)}
-                aria-label={playlist.description}
-                aria-live={index === activeIndex ? "assertive" : "off"}
-              >
-                <img src={playlist.imageUrl} alt={playlist.description} />
-              </div>
-            ))}
-          </Slider>
-        </div>
-        <input type='hidden' id={`playlist-announce-${id}`} className="sr-only" aria-live="assertive"></input>
-        <div className='carrusel-dot-section' />
-      </div>
-    </section>
-  );
-};
-
-export default CarouselComponent;
