+++ Dutify/src/components/locations/inicio/inicio.jsx
@@ -3,8 +3,7 @@ import { useEffect, useState } from "react";
 import { getUserPlaylists, getPopularPlaylists, getPopularArtistsPlaylists, getRecommendedPlaylists } from "../../../spotifyApi/SpotifyApiCalls";
 import CardsGrid from "../../cardsGrid/cardsGrid";
 import Spinner from '../../spinner/spinner';
-import CustomCarousel from '../../carousel/carousel';
-import CarouselComponent from '../../carousel/prueba';
+import CarouselComponent from '../../carousel/carousel';
 
 function Inicio({token}){
     
@@ -19,10 +18,10 @@ function Inicio({token}){
         
         setLoading(true);
         /*const user_playlists = await getUserPlaylists(token)
-        setRecentPlaylists(user_playlists.slice(0,6)) // PENDIENTE DE CAMBIO
+        setRecentPlaylists(user_playlists.slice(0,6)) // PENDIENTE DE CAMBIO*/
         setPopularPlaylists(await getPopularPlaylists(token))
-        setRecommendedPlaylists(await getRecommendedPlaylists(token))*/
-        setPopularArtistsPlaylists(await getPopularArtistsPlaylists(token))
+        /*setRecommendedPlaylists(await getRecommendedPlaylists(token))
+        setPopularArtistsPlaylists(await getPopularArtistsPlaylists(token))*/
     }
 
     useEffect(() => {
@@ -42,12 +41,9 @@ function Inicio({token}){
                 <h5 className='h5-recent-lists' tabIndex="0" aria-labelledby="Listas recientes">Listas recientes:</h5>
                 <CardsGrid type="genrelists" data={recent_playlists} clickFunction={listButtonClickHandler} />
             </div>
-            {/*<CarouselComponent id="Listas populares" lista={popular_playlists} name="Listas populares:"></CarouselComponent>
-            */}
+            <CarouselComponent id="Listas populares" lista={popular_playlists} name="Listas populares:"></CarouselComponent>
             <CarouselComponent id="Tus artistas favoritos" lista={popular_artists_playlists} name="Tus artistas favoritos:"></CarouselComponent>
-            {/*
             <CarouselComponent id="Recomendaciones" lista={recommended_playlists} name="Recomendaciones:"></CarouselComponent>
-            */}
             </>
             }
         </section>
